webhook/requirements.txt
flask==3.0.0
supabase==2.1.3
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
